~ | updated docker to compile new car control program, and moved car_cluster to correct directory
