--- src/Kuksa/car_cluster_with_kuksa/resources.qrc
@@ -1,7 +1,0 @@
-<RCC>
-    <qresource prefix="/car_cluster">
-        <file>Main.qml</file>
-        <file>mclaren.png</file>
-    </qresource>
-</RCC>
-
